packages/agent_toolkit/pyproject.toml
[project]
name = "agent-toolkit"
version = "0.1.0"
description = "Reusable agent runtime toolkit for the playground"
readme = "README.md"
authors = [
    { name = "Agentic AI Playground Contributors", email = "contributors@example.com" }
]
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.1.0,<1.0.0",
    "mcp>=1.0.0",
    "botocore[crt]>=1.34.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.0",
    "pydantic>=2.0.0",
    "strands-agents-evals>=0.1.3,<0.2.0",
    "arize-phoenix-otel>=0.9.0",
    "arize-phoenix-evals>=0.24.0",
    "litellm>=1.60.0",
    "openinference-instrumentation-bedrock>=0.1.30",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
